Заменил картинки для splash_screen на png. Сам SplashScreen сделан как фрагмент.
