--- app/src/main/java/com/bryukhanov/shoppinglist/MainActivity.kt
@@ -1,34 +1,0 @@
-package com.bryukhanov.shoppinglist
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.splash_screen)
-
-        CoroutineScope(Dispatchers.Main).launch {
-            delay(DELAY_SPLASH_SCREEN)
-            setContentView(R.layout.activity_main)
-
-            ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-                val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-                v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
-                insets
-            }
-        }
-    }
-
-    companion object {
-        private const val DELAY_SPLASH_SCREEN = 1500L
-    }
-}
