+++ app/src/main/java/com/bryukhanov/shoppinglist/core/ui/SplashScreenFragment.kt
@@ -1,0 +1,45 @@
+package com.bryukhanov.shoppinglist.core.ui
+
+import android.annotation.SuppressLint
+import android.content.pm.ActivityInfo
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import com.bryukhanov.shoppinglist.R
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
+
+@SuppressLint("CustomSplashScreen")
+class SplashScreenFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        return inflater.inflate(R.layout.fragment_splash_screen, container, false)
+    }
+
+    @SuppressLint("SourceLockedOrientationActivity")
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        requireActivity().requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+
+        viewLifecycleOwner.lifecycleScope.launch {
+            delay(DELAY_SPLASH_SCREEN)
+            findNavController().navigate(R.id.action_splashScreenFragment_to_myListsFragment)
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        requireActivity().requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED
+    }
+
+    companion object {
+        private const val DELAY_SPLASH_SCREEN = 1500L
+    }
+}
