+++ app/src/main/res/values/strings.xml
@@ -2,6 +2,6 @@
     <string name="app_name">ShoppingList</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="splash_screen_text1">Добро пожаловать в Список покупок!</string>
-    <string name="splash_screen_text2">Создавайте списки, добавляйте товары,\nотмечайте, что уже куплено</string>
+    <string name="splash_screen_text_welcome">Добро пожаловать в Список покупок!</string>
+    <string name="splash_screen_text_tasks">Создавайте списки, добавляйте товары,\nотмечайте, что уже куплено</string>
 </resources>
